extract to worker
