+++ src/backend/libs/Bridge.js
@@ -103,22 +103,6 @@ class Bridge extends Bot {
 
     }
   }
-
-  async getJob() {
-    return true;
-  }
-
-  async doJob() {
-    return true;
-  }
-
-  async finishJob() {
-    return true;
-  }
-
-  async updateJob() {
-    return true;
-  }
 }
 
 module.exports = Bridge;
